requirements.txt
alembic==1.7.7
allure-pytest==2.9.45
allure-python-commons==2.9.45
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
et-xmlfile==1.1.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.6.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-rerunfailures==10.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
test_name==1.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wsproto==1.1.0
zipp==3.6.0
